backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djongo==1.3.6
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
huggingface-hub==0.19.4
celery==5.3.4
redis==5.0.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
gunicorn==21.2.0
whitenoise==6.6.0